ipsec,tests: update key-rotation tests for VXLAN-in-ESP traffic
The key-rotation tests originally added a fixed integer to the number of
keys it waits for if Encrypted Overlay was enabled.

Update the tests to add this fixed integer to the keys to wait for if
tunnel mode is enabled.

Cilium now installs VXLAN-in-ESP states when tunnel mode is enabled and
not only when Encrypted Overlay is enabled.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>