ci: bpftrace skip tracking connections outside pod CIDRs
The current bpftrace script we use in CI to track and ensure that no
unencrypted packets are leaked does not check for the destination address
to belong to pod CIDRs. This might cause problems in egress-gateway tests,
since pod-to-outside and pod-to-node traffic is generated.

We currently avoid this by running egress-gateway related CLI tests
separately from all the others and with no leak-detection script running.
As we're going to introduce leak detection in more and more CI tests
hopefully, let's add this check to skip tracking unneeded traffic.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>