k8s: Debounce service events
[ upstream commit 82effe0eddbda0b98399c96d5f40ba221ee5d547 ]

[ Backporter's notes: Decreased number of expected deletes in
TestServiceEventDebounce, as https://github.com/cilium/cilium/pull/33434
has not been backported and ANY protocol is not expected. ]

When a service has multiple endpoint slices associated with it
this will cause Cilium to process each slice separately, causing
repeated updating of the service BPF maps and other churn.

To reduce the processing in these cases, debounce the processing
of the service events by collecting events into a buffer every 200ms.
If a new event arrives for a specific service during this duration it
is coalesced with the prior one.

Signed-off-by: Jussi Maki <jussi@isovalent.com>