clustermesh: Add hidden flag --allow-buggy-clustermesh
This hidden flag can allow existing users who are using a high
clustermesh ID in ENI environments to continue using it at their own
risk. Recently we added a check to startup to detect this potentially
problematic configuration and warn users, however it's now no longer
possible to run Cilium in the risky configuration. This puts users in a
tough spot of whether to run an outdated release or manually patch out
this newly introduced guardrail.

In the long term we should be able to lift the restriction around
usage of clustermesh ID in the datapath, but for now we should try to
encourage users not to configure Cilium this way while still ultimately
leaving the decision up to them about how to manage their environments.

Signed-off-by: Joe Stringer <joe@cilium.io>